Added local download when button is pressed
however because the file is not from the same origin the suggested file
name in download={suggested} does not work, this is related to the
content-disposition header and should be fixed. It is also possible to
use the .download method like in the original attempt but that does not
prompt the user for their preferred download location, so a UI element
would need to be added for them to enter the download location before
the download is initiated by the firebase api. the os npm package
(already included in file) may help